Adding a more-versatile condition-handling method for preventing integration tests from mashing on one another.

+++ django_mailbox/transports/generic.py
@@ -16,10 +16,12 @@ class GenericFileMailbox(EmailTransport):
     def get_instance(self):
         return self._variant(self._path)
 
-    def get_message(self):
+    def get_message(self, condition=None):
         repository = self.get_instance()
         repository.lock()
         for key, message in repository.items():
+            if condition and not condition(message):
+                continue
             repository.remove(key)
             yield message
         repository.unlock()
